// client/models.ts
export interface IExpansions {
  chronicleOfTheGodslayer: boolean;
  returnOfTheFallen: boolean;
  stormOfSouls: boolean;
  immortalHeroes: boolean;
  riseOfVigil: boolean;
  darknessUnleashed: boolean;
  realmsUnraveled: boolean;
  dawnOfChampions: boolean;
  dreamscape: boolean;
  warOfShadows: boolean;
  giftOfTheElements: boolean;
  valleyOfTheAncients: boolean;
  delerium: boolean;
  deliverance: boolean;
  promo1: boolean;
  promo2: boolean;
  promo3: boolean;
  promo4: boolean;
  promo5: boolean;
  promo6: boolean;
}

export type ExpansionObject = {
  name: string;
  code: string;
  grid: number;
}